+++ tests/testthat/test-rm_between.R
@@ -46,11 +46,22 @@ test_that("rm_between is extracting betweened text multiple markers",{
 		
 })
 
-test_that("rm_between is extracting betweened quotes",{
+test_that("rm_between is extracting between double quotes",{
     
     x <- 'Fresh or chilled Atlantic salmon "Salmo salar" and Danube salmon "Hucho hucho"'
     
     expect_equal(rm_between(x, '"', '"', extract = TRUE),
+        structure(list(c("Salmo salar", "Hucho hucho")), class = c("extracted", 
+            "list"))
+    )
+
+})
+
+test_that("rm_between is extracting between single quotes",{
+    
+    x <- "Fresh or chilled Atlantic salmon 'Salmo salar' and Danube salmon 'Hucho hucho'"
+    
+    expect_equal(rm_between(x, "'", "'", extract = TRUE),
         structure(list(c("Salmo salar", "Hucho hucho")), class = c("extracted", 
             "list"))
     )